Convert homu status updates and review assignment to handlers.
